# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Enterprise Content Platform Control API",
    "version": "1.0.0",
    "description": "C02 contract export. C03 implements the HTTP handlers."
  },
  "paths": {
...
    "/api/v1/me": {
      "get": {
        "operationId": "getCurrentIdentity",
        "responses": {
          "200": {
            "description": "Current identity and workspaces",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentIdentitySuccess"
                }
              }
            }
          },
          "403": {
            "description": "Workspace access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "503": {
            "description": "Identity unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workspaces": {
      "get": {
        "operationId": "listWorkspaces",
        "responses": {
          "200": {
            "description": "Accessible workspaces",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceListSuccess"
                }
              }
            }
          },
          "403": {
            "description": "Workspace access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects": {
      "get": {
        "operationId": "listProjects",
        "responses": {
          "200": {
            "description": "Projects in the current workspace",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectListSuccess"
                }
              }
            }
          },
          "403": {
            "description": "Workspace access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createProject",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectCommand"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectSuccess"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "403": {
            "description": "Workspace access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}": {
      "get": {
        "operationId": "getProject",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProjectId"
          }
        ],
        "responses": {
          "200": {
            "description": "Project, shots, and assets",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetailSuccess"
                }
              }
            }
          },
          "403": {
            "description": "Workspace access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateProject",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          },
          {
            "$ref": "#/components/parameters/ProjectId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Project updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectSuccess"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "403": {
            "description": "Workspace access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "409": {
            "description": "Idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 255
        }
      },
      "ProjectId": {
        "name": "project_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^prj_"
        }
      },
...
    },
    "schemas": {
...
      "ApiFailure": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "AUTH_UNAVAILABLE",
                  "AUTH_FORBIDDEN",
                  "CREDIT_INSUFFICIENT",
                  "CREDIT_CONFLICT",
                  "CREDIT_UNAVAILABLE",
                  "PROVIDER_UNAVAILABLE",
                  "PROVIDER_REJECTED",
                  "PROVIDER_PROTOCOL_INVALID",
                  "DOWNLOAD_INVALID",
                  "IDEMPOTENCY_CONFLICT",
                  "WORKSPACE_FORBIDDEN",
                  "VALIDATION_FAILED",
                  "NOT_FOUND",
                  "TASK_STATE_INVALID",
                  "INTERNAL_ERROR"
                ]
              },
              "message": {
                "type": "string",
                "minLength": 1
              },
              "retryable": {
                "type": "boolean"
              },
              "details": {
                "type": "object",
                "additionalProperties": {},
                "default": {}
              }
            },
            "required": [
              "code",
              "message",
              "retryable"
            ],
            "additionalProperties": false
          },
          "request_id": {
            "type": "string",
            "pattern": "^req_[0-9A-HJKMNP-TV-Z]{26}$"
          }
        },
        "required": [
          "error",
          "request_id"
        ],
        "additionalProperties": false
      },
...
      "ProjectSuccess": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "pattern": "^prj_[A-Za-z0-9][A-Za-z0-9_-]{2,63}$"
              },
              "workspace_id": {
                "type": "string",
                "pattern": "^ws_[A-Za-z0-9][A-Za-z0-9_-]{2,63}$"
              },
              "name": {
                "type": "string",
                "minLength": 1,
                "maxLength": 255
              },
              "status": {
                "type": "string",
                "enum": [
                  "ACTIVE",
                  "ARCHIVED"
                ]
              },
              "created_at": {
                "type": "string",
                "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
              },
              "updated_at": {
                "$ref": "#/definitions/ProjectSuccess/properties/data/properties/created_at"
              }
            },
            "required": [
              "id",
              "workspace_id",
              "name",
              "status",
              "created_at",
              "updated_at"
            ],
            "additionalProperties": false
          },
          "request_id": {
            "type": "string",
            "pattern": "^req_[0-9A-HJKMNP-TV-Z]{26}$"
          }
        },
        "required": [
          "data",
          "request_id"
        ],
        "additionalProperties": false
      },
      "ProjectListSuccess": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "pattern": "^prj_[A-Za-z0-9][A-Za-z0-9_-]{2,63}$"
                },
                "workspace_id": {
                  "type": "string",
                  "pattern": "^ws_[A-Za-z0-9][A-Za-z0-9_-]{2,63}$"
                },
                "name": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 255
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "ACTIVE",
                    "ARCHIVED"
                  ]
                },
                "created_at": {
                  "type": "string",
                  "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
                },
                "updated_at": {
                  "$ref": "#/definitions/ProjectListSuccess/properties/data/items/properties/created_at"
                }
              },
              "required": [
                "id",
                "workspace_id",
                "name",
                "status",
                "created_at",
                "updated_at"
              ],
              "additionalProperties": false
            }
          },
          "request_id": {
            "type": "string",
            "pattern": "^req_[0-9A-HJKMNP-TV-Z]{26}$"
          }
        },
        "required": [
          "data",
          "request_id"
        ],
        "additionalProperties": false
      },
...
      "CurrentIdentitySuccess": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "user": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "pattern": "^usr_[A-Za-z0-9][A-Za-z0-9_-]{2,63}$"
                  },
                  "display_name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "ACTIVE",
                      "DISABLED"
                    ]
                  },
                  "created_at": {
                    "type": "string",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
                  },
                  "updated_at": {
                    "$ref": "#/definitions/CurrentIdentitySuccess/properties/data/properties/user/properties/created_at"
                  }
                },
                "required": [
                  "id",
                  "display_name",
                  "status",
                  "created_at",
                  "updated_at"
                ],
                "additionalProperties": false
              },
              "workspaces": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "pattern": "^ws_[A-Za-z0-9][A-Za-z0-9_-]{2,63}$"
                    },
                    "name": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 255
                    },
                    "created_by": {
                      "$ref": "#/definitions/CurrentIdentitySuccess/properties/data/properties/user/properties/id"
                    },
                    "created_at": {
                      "$ref": "#/definitions/CurrentIdentitySuccess/properties/data/properties/user/properties/created_at"
                    },
                    "updated_at": {
                      "$ref": "#/definitions/CurrentIdentitySuccess/properties/data/properties/user/properties/created_at"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "created_by",
                    "created_at",
                    "updated_at"
                  ],
                  "additionalProperties": false
                }
              }
            },
            "required": [
              "user",
              "workspaces"
            ],
            "additionalProperties": false
          },
          "request_id": {
            "type": "string",
            "pattern": "^req_[0-9A-HJKMNP-TV-Z]{26}$"
          }
        },
        "required": [
          "data",
          "request_id"
        ],
        "additionalProperties": false
      },
...
      "WorkspaceListSuccess": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "pattern": "^ws_[A-Za-z0-9][A-Za-z0-9_-]{2,63}$"
                },
                "name": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 255
                },
                "created_by": {
                  "type": "string",
                  "pattern": "^usr_[A-Za-z0-9][A-Za-z0-9_-]{2,63}$"
                },
                "created_at": {
                  "type": "string",
                  "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
                },
                "updated_at": {
                  "$ref": "#/definitions/WorkspaceListSuccess/properties/data/items/properties/created_at"
                }
              },
              "required": [
                "id",
                "name",
                "created_by",
                "created_at",
                "updated_at"
              ],
              "additionalProperties": false
            }
          },
          "request_id": {
            "type": "string",
            "pattern": "^req_[0-9A-HJKMNP-TV-Z]{26}$"
          }
        },
        "required": [
          "data",
          "request_id"
        ],
        "additionalProperties": false
      },
      "CreateProjectCommand": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          }
        },
        "required": [
          "name"
        ],
        "additionalProperties": false
      },
      "UpdateProjectCommand": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "ARCHIVED"
            ]
          }
        },
        "additionalProperties": false
      },
...
      "ProjectDetailSuccess": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "project": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "pattern": "^prj_[A-Za-z0-9][A-Za-z0-9_-]{2,63}$"
                  },
                  "workspace_id": {
                    "type": "string",
                    "pattern": "^ws_[A-Za-z0-9][A-Za-z0-9_-]{2,63}$"
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "ACTIVE",
                      "ARCHIVED"
                    ]
                  },
                  "created_at": {
                    "type": "string",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
                  },
                  "updated_at": {
                    "$ref": "#/definitions/ProjectDetailSuccess/properties/data/properties/project/properties/created_at"
                  }
                },
                "required": [
                  "id",
                  "workspace_id",
                  "name",
                  "status",
                  "created_at",
                  "updated_at"
                ],
                "additionalProperties": false
              },
              "shots": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "pattern": "^sht_[A-Za-z0-9][A-Za-z0-9_-]{2,63}$"
                    },
                    "workspace_id": {
                      "$ref": "#/definitions/ProjectDetailSuccess/properties/data/properties/project/properties/workspace_id"
                    },
                    "project_id": {
                      "$ref": "#/definitions/ProjectDetailSuccess/properties/data/properties/project/properties/id"
                    },
                    "position": {
                      "type": "integer",
                      "minimum": 0
                    },
                    "prompt": {
                      "type": "string"
                    },
                    "model": {
                      "type": "string",
                      "minLength": 1,
                      "nullable": true
                    },
                    "generation_settings": {
                      "type": "object",
                      "additionalProperties": {}
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "DRAFT",
                        "READY",
                        "GENERATING",
                        "GENERATED",
                        "FAILED",
                        "ARCHIVED"
                      ]
                    },
                    "selected_asset_id": {
                      "type": "string",
                      "pattern": "^ast_[A-Za-z0-9][A-Za-z0-9_-]{2,63}$",
                      "nullable": true
                    },
                    "revision": {
                      "type": "integer",
                      "exclusiveMinimum": true,
                      "minimum": 0
                    },
                    "created_at": {
                      "$ref": "#/definitions/ProjectDetailSuccess/properties/data/properties/project/properties/created_at"
                    },
                    "updated_at": {
                      "$ref": "#/definitions/ProjectDetailSuccess/properties/data/properties/project/properties/created_at"
                    }
                  },
                  "required": [
                    "id",
                    "workspace_id",
                    "project_id",
                    "position",
                    "prompt",
                    "model",
                    "generation_settings",
                    "status",
                    "selected_asset_id",
                    "revision",
                    "created_at",
                    "updated_at"
                  ],
                  "additionalProperties": false
                }
              },
              "assets": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "$ref": "#/definitions/ProjectDetailSuccess/properties/data/properties/shots/items/properties/selected_asset_id"
                    },
                    "workspace_id": {
                      "$ref": "#/definitions/ProjectDetailSuccess/properties/data/properties/project/properties/workspace_id"
                    },
                    "project_id": {
                      "$ref": "#/definitions/ProjectDetailSuccess/properties/data/properties/project/properties/id"
                    },
                    "kind": {
                      "type": "string",
                      "enum": [
                        "IMAGE",
                        "VIDEO",
                        "AUDIO",
                        "DOCUMENT",
                        "POSTER",
                        "THUMBNAIL"
                      ]
                    },
                    "origin": {
                      "type": "string",
                      "enum": [
                        "USER_UPLOAD",
                        "GENERATED",
                        "DERIVED"
                      ]
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "PENDING_UPLOAD",
                        "READY",
                        "FAILED",
                        "DELETED"
                      ]
                    },
                    "sha256": {
                      "type": "string",
                      "pattern": "^[a-f0-9]{64}$",
                      "nullable": true
                    },
                    "mime_type": {
                      "type": "string",
                      "minLength": 1,
                      "nullable": true
                    },
                    "byte_size": {
                      "type": "integer",
                      "minimum": 0,
                      "nullable": true
                    },
                    "width": {
                      "type": "integer",
                      "exclusiveMinimum": true,
                      "minimum": 0,
                      "nullable": true
                    },
                    "height": {
                      "type": "integer",
                      "exclusiveMinimum": true,
                      "minimum": 0,
                      "nullable": true
                    },
                    "duration_ms": {
                      "type": "integer",
                      "minimum": 0,
                      "nullable": true
                    },
                    "metadata": {
                      "$ref": "#/definitions/ProjectDetailSuccess/properties/data/properties/shots/items/properties/generation_settings"
                    },
                    "created_at": {
                      "$ref": "#/definitions/ProjectDetailSuccess/properties/data/properties/project/properties/created_at"
                    },
                    "updated_at": {
                      "$ref": "#/definitions/ProjectDetailSuccess/properties/data/properties/project/properties/created_at"
                    }
                  },
                  "required": [
                    "id",
                    "workspace_id",
                    "project_id",
                    "kind",
                    "origin",
                    "status",
                    "sha256",
                    "mime_type",
                    "byte_size",
                    "width",
                    "height",
                    "duration_ms",
                    "metadata",
                    "created_at",
                    "updated_at"
                  ],
                  "additionalProperties": false
                }
              }
            },
            "required": [
              "project",
              "shots",
              "assets"
            ],
            "additionalProperties": false
          },
          "request_id": {
            "type": "string",
            "pattern": "^req_[0-9A-HJKMNP-TV-Z]{26}$"
          }
        },
        "required": [
          "data",
          "request_id"
        ],
        "additionalProperties": false
      },
...
    }
  }
}